requirements.txt
flask==2.1.0
pandas==1.4.2
scikit-learn==1.0.2
gunicorn==20.1.0
python-dotenv==1.0.0